Move ConnectionManagerController to Stratis.Bitcoin.Connection
